+++ shared/config/config_test.go
@@ -5,7 +5,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/xmlking/micro-starter-kit/shared/config"
+    "google.golang.org/grpc/resolver"
+
+    "github.com/xmlking/micro-starter-kit/shared/config"
 )
 
 // CONFIGOR_DEBUG_MODE=true go test -v ./shared/config/... -count=1
@@ -45,4 +47,45 @@ func ExampleGetConfig_check_defaults() {
 	// mkit.service.account:8080
 	// v0.1.0
 	// 8888
+}
+
+func TestParseTargetString(t *testing.T) {
+    for _, test := range []struct {
+        targetStr string
+        want      resolver.Target
+    }{
+        {targetStr: "", want: resolver.Target{Scheme: "", Authority: "", Endpoint: ""}},
+        {targetStr: ":///", want: resolver.Target{Scheme: "", Authority: "", Endpoint: ""}},
+        {targetStr: "a:///", want: resolver.Target{Scheme: "a", Authority: "", Endpoint: ""}},
+        {targetStr: "://a/", want: resolver.Target{Scheme: "", Authority: "a", Endpoint: ""}},
+        {targetStr: ":///a", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "a"}},
+        {targetStr: "a://b/", want: resolver.Target{Scheme: "a", Authority: "b", Endpoint: ""}},
+        {targetStr: "a:///b", want: resolver.Target{Scheme: "a", Authority: "", Endpoint: "b"}},
+        {targetStr: "://a/b", want: resolver.Target{Scheme: "", Authority: "a", Endpoint: "b"}},
+        {targetStr: "a://b/c", want: resolver.Target{Scheme: "a", Authority: "b", Endpoint: "c"}},
+        {targetStr: "dns:///google.com", want: resolver.Target{Scheme: "dns", Authority: "", Endpoint: "google.com"}},
+        {targetStr: "dns:///google.com:8080", want: resolver.Target{Scheme: "dns", Authority: "", Endpoint: "google.com:8080"}},
+        {targetStr: "dns://a.server.com/google.com", want: resolver.Target{Scheme: "dns", Authority: "a.server.com", Endpoint: "google.com"}},
+        {targetStr: "dns://a.server.com/google.com/?a=b", want: resolver.Target{Scheme: "dns", Authority: "a.server.com", Endpoint: "google.com/?a=b"}},
+
+        {targetStr: "/", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "/"}},
+        {targetStr: "google.com", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "google.com"}},
+        {targetStr: "google.com/?a=b", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "google.com/?a=b"}},
+        {targetStr: "/unix/socket/address", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "/unix/socket/address"}},
+        {targetStr: "unix:///tmp/mysrv.sock", want: resolver.Target{Scheme: "unix", Authority: "", Endpoint: "tmp/mysrv.sock"}},
+
+        // If we can only parse part of the target.
+        {targetStr: "://", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "://"}},
+        {targetStr: "unix://domain", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "unix://domain"}},
+        {targetStr: "a:b", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "a:b"}},
+        {targetStr: "a/b", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "a/b"}},
+        {targetStr: "a:/b", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "a:/b"}},
+        {targetStr: "a//b", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "a//b"}},
+        {targetStr: "a://b", want: resolver.Target{Scheme: "", Authority: "", Endpoint: "a://b"}},
+    } {
+        got := config.ParseTarget(test.targetStr)
+        if got != test.want {
+            t.Errorf("ParseTarget(%q) = %+v, want %+v", test.targetStr, got, test.want)
+        }
+    }
 }