+++ shared/constants/constants.go
@@ -1,6 +1,7 @@
 package constants
 
 const (
+    ACCOUNT_CLIENT   = "mkit.client.account"
 	ACCOUNT_SERVICE  = "mkit.service.account"
 	GREETER_SERVICE  = "mkit.service.greeter"
 	EMAILER_SERVICE  = "mkit.service.emailer"