+++ shared/constants/metadata.go
@@ -4,6 +4,8 @@ package constants
 
 const (
 	// There are certain requirements for metadata to be passed in the http header:
-	// Just use all capital letters for God's sake
-	TransID = "TRANS-ID"
+	// gRPC recommended Key format: `lowercase alphanumeric characters and hyphen`
+	// but go-micro use `camelcase alphanumeric characters and hyphen`
+    TraceIDKey = "Micro-Trace-Id"
+    TenantIdKey = "Tenant-Id"
 )