+++ shared/wrapper/transaction/transaction.wrapper.go
@@ -1,18 +1,15 @@
 package transaction
 
 import (
-	"context"
+    "context"
 
-	"github.com/golang/protobuf/proto"
-	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/google/uuid"
-	"github.com/micro/go-micro/v2"
-	"github.com/micro/go-micro/v2/metadata"
-	"github.com/micro/go-micro/v2/server"
-	"github.com/rs/zerolog/log"
+    "github.com/golang/protobuf/proto"
+    "github.com/golang/protobuf/ptypes/empty"
+    "github.com/micro/go-micro/v2"
+    "github.com/micro/go-micro/v2/server"
+    "github.com/rs/zerolog/log"
 
-	transactionPB "github.com/xmlking/micro-starter-kit/service/recorder/proto/transaction"
-	"github.com/xmlking/micro-starter-kit/shared/constants"
+    transactionPB "github.com/xmlking/micro-starter-kit/service/recorder/proto/transaction"
 )
 
 func publish(ctx context.Context, publisher micro.Event, req, rsp proto.Message) (err error) {
@@ -37,9 +34,6 @@ func publish(ctx context.Context, publisher micro.Event, req, rsp proto.Message)
 func NewHandlerWrapper(p micro.Event) server.HandlerWrapper {
 	return func(fn server.HandlerFunc) server.HandlerFunc {
 		return func(ctx context.Context, req server.Request, rsp interface{}) (err error) {
-			// add TranID to context if not present
-			// ctx = metadata.Set(ctx, constants.TransID, uuid.New().String())
-			ctx = metadata.MergeContext(ctx, map[string]string{constants.TransID: uuid.New().String()}, false)
 			err = fn(ctx, req, rsp)
 			// we already logged error in Publish. lets ignore error here. # Note: this is blocking call..
 			_ = publish(ctx, p, req.Body().(proto.Message), rsp.(proto.Message))
@@ -53,8 +47,6 @@ func NewHandlerWrapper(p micro.Event) server.HandlerWrapper {
 func NewSubscriberWrapper(p micro.Event) server.SubscriberWrapper {
 	return func(fn server.SubscriberFunc) server.SubscriberFunc {
 		return func(ctx context.Context, req server.Message) (err error) {
-			// add TranID to context if not present
-			ctx = metadata.MergeContext(ctx, map[string]string{constants.TransID: uuid.New().String()}, false)
 			err = fn(ctx, req)
 			// we already logged error in Publish. lets ignore error here.
 			// FIXME: `Micro-From-Service` is not replaced